+++ src/FileCollection.ts
@@ -1,0 +1,3 @@
+import { FileItem } from './FileItem';
+
+export type FileCollection = FileItem[];