+++ src/PartialFile.ts
@@ -1,4 +1,0 @@
-/** subset for File */
-
-export type FileItem = Omit<File, 'slice' | 'type'>;
-export type FileItemList = FileItem[];