+++ src/index.ts
@@ -4,4 +4,5 @@ export * from './fileCollectionFromWebservice';
 export * from './fileCollectionUnzip';
 export * from './fileCollectionUngzip';
 export * from './groupFiles';
+export * from './FileCollection';
 export * from './FileItem';